meltano/requirements.txt
pandas
pyarrow
psycopg2-binary
sqlalchemy